@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import contextlib
 import time
 import unittest
@@ -9,8 +7,6 @@ from concurrent.futures.thread import ThreadPoolExecutor
 from datetime import date, datetime
 from decimal import Decimal
 from random import randint
-
-from past.builtins.misc import xrange
 
 from pyathenajdbc import BINARY, BOOLEAN, DATE, DATETIME, NUMBER, STRING, connect
 from pyathenajdbc.cursor import Cursor
@@ -40,7 +36,7 @@ class TestCursor(unittest.TestCase, WithConnect):
         cursor.execute("SELECT * FROM one_row")
         self.assertEqual(cursor.fetchall(), [(1,)])
         cursor.execute("SELECT a FROM many_rows ORDER BY a")
-        self.assertEqual(cursor.fetchall(), [(i,) for i in xrange(10000)])
+        self.assertEqual(cursor.fetchall(), [(i,) for i in range(10000)])
 
     @with_cursor
     def test_null_param(self, cursor):
@@ -232,7 +228,7 @@ class TestCursor(unittest.TestCase, WithConnect):
         self.assertEqual(cursor.fetchall(), [(None,)] * 10000)
         cursor.execute("SELECT IF(a % 11 = 0, null, a) FROM many_rows")
         self.assertEqual(
-            cursor.fetchall(), [(None if a % 11 == 0 else a,) for a in xrange(10000)]
+            cursor.fetchall(), [(None if a % 11 == 0 else a,) for a in range(10000)]
         )
 
     @with_cursor
@@ -418,10 +414,10 @@ class TestCursor(unittest.TestCase, WithConnect):
     def test_executemany(self, cursor):
         cursor.executemany(
             "INSERT INTO execute_many (a) VALUES (%(a)s)",
-            [{"a": i} for i in xrange(1, 3)],
+            [{"a": i} for i in range(1, 3)],
         )
         cursor.execute("SELECT * FROM execute_many")
-        self.assertEqual(sorted(cursor.fetchall()), [(i,) for i in xrange(1, 3)])
+        self.assertEqual(sorted(cursor.fetchall()), [(i,) for i in range(1, 3)])
 
     @with_cursor
     def test_executemany_fetch(self, cursor):
